LayoutTests/http/tests/xmlhttprequest/access-control-basic-post-fail-non-simple-content-type-expected.txt

CONSOLE MESSAGE: line 28: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-options-not-supported.cgi. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [NetworkError: DOM Exception 19].
